alloc::boxed: enable test
Previously test was disabled due to `#[cfg(test)]` before `mod boxed`.
